# Embeddings: add an AVX512 implementation of Dot