--- a/enterprise/internal/embeddings/dot_amd64.s
+++ b/enterprise/internal/embeddings/dot_amd64.s
@@ -74,3 +74,71 @@ end:
 	MOVQ R8, ret+48(FP)
 	RET
 
+// dotVNNI calculates the dot product of two slices using AVX512 VNNI
+// instructions The slices must be of equal length and that length must be a
+// multiple of 64.
+TEXT ·dotVNNI(SB), NOSPLIT, $0-52
+	// Offsets based on slice header offsets.
+	// To check, use `GOARCH=amd64 go vet`
+	MOVQ a_base+0(FP), AX
+	MOVQ b_base+24(FP), BX
+	MOVQ a_len+8(FP), DX
+
+    ADDQ AX, DX // end pointer
+
+	// Zero our accumulators
+	VPXORQ Z0, Z0, Z0 // positive
+	VPXORQ Z1, Z1, Z1 // negative
+
+	// Fill Z3 with 128
+	MOVD $0x80808080, R9
+	VPBROADCASTD R9, Z2
+
+blockloop:
+	CMPQ AX, DX
+	JE reduce
+
+	VMOVDQU8 (AX), Z3
+	VMOVDQU8 (BX), Z4
+
+	// The VPDPBUSD instruction calculates of the dot product 4 columns at a
+	// time, accumulating into an i32 vector. The problem is it expects one
+	// vector to be unsigned bytes and one to be signed bytes. To make this
+	// work, we make one of our vectors unsigned by adding 128 to each element.
+	// This causes us to overshoot, so we keep track of the amount we need
+	// to compensate by so we can subtract it from the sum at the end.
+	//
+	// Effectively, we are calculating SUM((Z3 + 128) · Z4) - 128 * SUM(Z4).
+
+	VPADDB Z3, Z2, Z3   // add 128 to Z3, making it unsigned
+	VPDPBUSD Z4, Z3, Z0 // Z0 += Z3 dot Z4
+	VPDPBUSD Z4, Z2, Z1 // Z1 += broadcast(128) dot Z4
+
+	ADDQ $64, AX
+	ADDQ $64, BX
+	JMP blockloop
+
+reduce:
+	VPSUBD Z1, Z0, Z0 // Z0 -= Z1
+
+    // Sum Z0 horizontally
+    VEXTRACTI64X4 $1, Z0, Y1
+    VPADDD Y0, Y1, Y0
+
+	VEXTRACTI128 $1, Y0, X1
+	VPADDD X0, X1, X0
+
+	VPSRLDQ $8, X0, X1
+	VPADDD X0, X1, X0
+
+	VPSRLDQ $4, X0, X1
+	VPADDD X0, X1, X0
+
+	// Store the reduced sum
+	VMOVD X0, R8
+
+end:
+	MOVL R8, ret+48(FP)
+	VZEROALL
+	RET
+