This adds a dot product implementation that uses [AVX512 VNNI instructions](https://iq.opengenus.org/avx512-vnni/). These instructions specifically exist to accelerate ML workloads. This yields a ~30% improvement in embeddings/s compared to the AVX2 implementation.

In the second commit, I cleaned up and simplified the implementations a bit since we now have 4 different implementations of the dot product (native go, arm64, amd64 with AVX2, and amd64 with AVX512-VNNI 🤯). Some comments inline for each of the changes.

```
goos: linux
goarch: amd64
pkg: github.com/sourcegraph/sourcegraph/enterprise/internal/embeddings
cpu: Intel(R) Xeon(R) Platinum 8481C CPU @ 2.70GHz
                                  │ /tmp/avx2.txt │            /tmp/vnni.txt            │
                                  │    sec/op     │   sec/op     vs base                │
SimilaritySearch/numWorkers=1-44      150.5m ± 3%   119.4m ± 1%  -20.67% (p=0.000 n=10)
SimilaritySearch/numWorkers=2-44      81.08m ± 6%   61.66m ± 1%  -23.95% (p=0.000 n=10)
SimilaritySearch/numWorkers=4-44      45.54m ± 2%   33.94m ± 4%  -25.48% (p=0.000 n=10)
SimilaritySearch/numWorkers=8-44      26.02m ± 1%   19.22m ± 3%  -26.15% (p=0.000 n=10)
SimilaritySearch/numWorkers=16-44     17.05m ± 0%   14.08m ± 2%  -17.42% (p=0.000 n=10)
geomean                               47.69m        36.81m       -22.80%

                                  │ /tmp/avx2.txt │            /tmp/vnni.txt             │
                                  │ embeddings/s  │ embeddings/s  vs base                │
SimilaritySearch/numWorkers=1-44      6.644M ± 3%    8.375M ± 1%  +26.05% (p=0.000 n=10)
SimilaritySearch/numWorkers=2-44      12.33M ± 5%    16.22M ± 1%  +31.49% (p=0.000 n=10)
SimilaritySearch/numWorkers=4-44      21.96M ± 2%    29.47M ± 4%  +34.18% (p=0.000 n=10)
SimilaritySearch/numWorkers=8-44      38.43M ± 1%    52.04M ± 3%  +35.40% (p=0.000 n=10)
SimilaritySearch/numWorkers=16-44     58.65M ± 0%    71.03M ± 2%  +21.10% (p=0.000 n=10)
geomean                               20.97M         27.16M       +29.53%

                                  │    /tmp/avx2.txt    │                /tmp/vnni.txt                │
                                  │ embeddings/s/worker │ embeddings/s/worker  vs base                │
SimilaritySearch/numWorkers=1-44            6.644M ± 3%           8.375M ± 1%  +26.05% (p=0.000 n=10)
SimilaritySearch/numWorkers=2-44            6.167M ± 5%           8.109M ± 1%  +31.49% (p=0.000 n=10)
SimilaritySearch/numWorkers=4-44            5.490M ± 2%           7.366M ± 4%  +34.18% (p=0.000 n=10)
SimilaritySearch/numWorkers=8-44            4.804M ± 1%           6.504M ± 3%  +35.40% (p=0.000 n=10)
SimilaritySearch/numWorkers=16-44           3.666M ± 0%           4.439M ± 2%  +21.10% (p=0.000 n=10)
geomean                                     5.243M                6.791M       +29.53%
```

## Test plan

Updated tests to run on cases larger than 64, add an additional "random ints" test to augment the more opaque quicktests. Manually ran tests and benchmarks on a GCE instance with AVX512-VNNI available. This is not enabled by default, and can be disabled easily by unsetting an environment variable, so the risk is low.

<!-- All pull requests REQUIRE a test plan: https://docs.sourcegraph.com/dev/background-information/testing_principles -->