@eseliger 2023-05-16
today I learned my raptor lake CPU is not fancy enough for this kind of stuff 😢