@camdencheek 2023-05-17
Yeah, it's pretty much only server-oriented CPUs that have this set. Thankfully, that includes all the GCE instances I tested on